reorganized files again into respective sections
